{{cookiecutter.package_name}}/requirements.txt
flask >= 1.0
python-dotenv
